Created KNN model
I made a rough KNN model based on our training and testing sets, but I have not tested it against the test set because we should make sure we are fully confident with everything else before we do that. I also made a graph showing experience if we want it. We maybe need to address the uneven distribution and whether we can do anything about it